Add files via upload
The google finance webpage (https://www.google.com/finance) reports a sector change summary in an HTML table at the bottom of the page. 
The table contains, amongst other things, the name of the sector, a link to a sector detail page, and the percentage change for that sector over the course of the day (the change is updated in real time).

Each sector detail page (e.g., https://www.google.com/finance?catid=TRBC%3A50&ei=erHdV6HdFcKgeJXHlrgO) contains specific information about that sector including a table titled ‘Top movers’. This table contains a list of stocks in the sector that have moved the most in percentage terms. The list contains the top five gainers, the top five losers, and the top five most active stocks (for the most recent day).
Write a function google_sector_report() that returns a json dump containing the following information about each sector:
1. The sector name
2. The percentage change in sector value
3. The biggest gainer and the percentage change in the b