Add files via upload
Your hedge fund boss decides that the fund needs a repository of sector changes information and appoints you as the person responsible for implementing the design of the repository, the collection of data, and the interface with the repository. You have three specific tasks that you need to do:
1. Write a program that grabs sector change data from google finance at the end of each day. Your boss doesn’t know this but, luckily, you’ve already written this for your data analytics class! Start collecting daily data by running this function after the market close every day and saving the JSON object somewhere - in a file perhaps.
2. Design a MongoDB collection that stores the data and write a function that updates the collection as new data arrives. You have three choices for the document design and whichever you think is the most appropriate or the easiest. Your choice will impact the way you answer the questions in part 3, so be careful!:
A. Each document represents one date. In thi